fixbug(main.yml infraestructure/[postal_codes_controller file_uploads_controller]): controller files were moved to infraestructure folder
