+++ README.md
@@ -10,6 +10,5 @@ Go to <a href="http://127.0.0.1:8000/docs" class="external-link" target="_blank"
  
 ### * Run project using manual commands
 ```
-pip install -r requirements.txt
 uvicorn main:app
 ```
